+++ src/command_modules/azure-cli-redis/azure/cli/command_modules/redis/_params.py
@@ -4,7 +4,6 @@
 # --------------------------------------------------------------------------------------------
 
 # pylint: disable=line-too-long
-from enum import Enum
 from azure.cli.core.commands.parameters import (
     get_resource_name_completion_list,
     enum_choice_list,
@@ -32,19 +31,6 @@ class JsonString(dict):
         dictval = shell_safe_json_parse(value)
         self.update(dictval)
 
-class VmSize(Enum): # pylint: disable=too-few-public-methods
-    C0 = 'C0'
-    C1 = 'C1'
-    C2 = 'C2'
-    C3 = 'C3'
-    C4 = 'C4'
-    C5 = 'C5'
-    C6 = 'C6'
-    P1 = 'P1'
-    P2 = 'P2'
-    P3 = 'P3'
-    P4 = 'P4'
-
 class ScheduleEntryList(list):
     def __init__(self, value):
         super(ScheduleEntryList, self).__init__()
@@ -65,10 +51,10 @@ register_cli_argument('redis', 'reboot_type', **enum_choice_list(RebootType))
 register_cli_argument('redis', 'key_type', **enum_choice_list(RedisKeyType))
 register_cli_argument('redis', 'shard_id', type=int)
 register_cli_argument('redis', 'sku', **enum_choice_list(SkuName))
-register_cli_argument('redis', 'vm_size', **enum_choice_list(VmSize))
+register_cli_argument('redis', 'vm_size', help='Size of redis cache to deploy. Example : values for C family (C0, C1, C2, C3, C4, C5, C6). For P family (P1, P2, P3, P4)')
 register_cli_argument('redis', 'enable_non_ssl_port', action='store_true')
 register_cli_argument('redis', 'shard_count', type=int)
-register_cli_argument('redis', 'subnet_id') # TODO: Create generic id completer similar to name
+register_cli_argument('redis', 'subnet_id')
 
 register_cli_argument('redis import-method', 'files', nargs='+')