+++ src/command_modules/azure-cli-redis/HISTORY.rst
@@ -4,7 +4,7 @@ Release History
 ===============
 
 * Adding update command which also adds the ability to scale for redis cache
-* Adding 'getting deprecated' message for update-settings command
+* Deprecates the 'update-settings' command.
 
 0.1.1b3 (2017-02-22)
 ++++++++++++++++++++